+++ pyexasol/__init__.py
@@ -27,6 +27,8 @@ __all__ = [
     'ExaLocalConfig',
     'HTTP_EXPORT',
     'HTTP_IMPORT',
+    'PROTOCOL_V1',
+    'PROTOCOL_V2',
 ]
 
 from .version import __version__
@@ -40,6 +42,7 @@ from .ext import ExaExtension
 from .meta import ExaMetaData
 from .mapper import exasol_mapper
 from .http_transport import ExaHTTPTransportWrapper, HTTP_EXPORT, HTTP_IMPORT
+from .constant import PROTOCOL_V1, PROTOCOL_V2
 
 
 def connect(**kwargs) -> ExaConnection: