+++ pyexasol/meta.py
@@ -1,3 +1,5 @@
+from . import constant
+from .exceptions import ExaRuntimeError
 
 
 class ExaMetaData(object):
@@ -6,7 +8,11 @@ class ExaMetaData(object):
     https://www.exasol.com/support/browse/IDEA-476
 
     If you still get locks, please make sure to update Exasol server to the latest minor version
+
+    This class also implements no SQL metadata commands introduced in Exasol v7.0 via .execute_meta_nosql() function
     """
+    snapshot_execution_hint = '/*snapshot execution*/'
+
     def __init__(self, connection):
         self.connection = connection
         self.sql_keywords = None
@@ -24,13 +30,18 @@ class ExaMetaData(object):
     def schema_exists(self, schema_name):
         object_name = self.connection.format.default_format_ident_value(schema_name)
 
-        st = self._execute_snapshot("""
-            SELECT 1
-            FROM sys.exa_schemas
-            WHERE schema_name={object_name}
-        """, {
-            'object_name': object_name,
-        })
+        if self.connection.protocol_version() >= constant.PROTOCOL_V2:
+            st = self.execute_meta_nosql('getSchemas', {
+                'schema': object_name,
+            })
+        else:
+            st = self.execute_snapshot("""
+                SELECT 1
+                FROM sys.exa_schemas
+                WHERE schema_name={object_name}
+            """, {
+                'object_name': object_name,
+            })
 
         return st.rowcount() > 0
 
@@ -42,15 +53,22 @@ class ExaMetaData(object):
             object_schema = self.connection.current_schema()
             object_name = self.connection.format.default_format_ident_value(table_name)
 
-        st = self._execute_snapshot("""
-            SELECT 1
-            FROM sys.exa_all_tables
-            WHERE table_schema={object_schema}
-                AND table_name={object_name}
-        """, {
-            'object_schema': object_schema,
-            'object_name': object_name,
-        })
+        if self.connection.protocol_version() >= constant.PROTOCOL_V2:
+            st = self.execute_meta_nosql('getTables', {
+                'schema': object_schema,
+                'table': object_name,
+                'tableTypes': ['TABLE'],
+            })
+        else:
+            st = self.execute_snapshot("""
+                SELECT 1
+                FROM sys.exa_all_tables
+                WHERE table_schema={object_schema}
+                    AND table_name={object_name}
+            """, {
+                'object_schema': object_schema,
+                'object_name': object_name,
+            })
 
         return st.rowcount() > 0
 
@@ -62,20 +80,27 @@ class ExaMetaData(object):
             object_schema = self.connection.current_schema()
             object_name = self.connection.format.default_format_ident_value(view_name)
 
-        st = self._execute_snapshot("""
-            SELECT 1
-            FROM sys.exa_all_views
-            WHERE view_schema={object_schema}
-                AND view_name={object_name}
-        """, {
-            'object_schema': object_schema,
-            'object_name': object_name,
-        })
+        if self.connection.protocol_version() >= constant.PROTOCOL_V2:
+            st = self.execute_meta_nosql('getTables', {
+                'schema': object_schema,
+                'table': object_name,
+                'tableTypes': ['VIEW'],
+            })
+        else:
+            st = self.execute_snapshot("""
+                SELECT 1
+                FROM sys.exa_all_views
+                WHERE view_schema={object_schema}
+                    AND view_name={object_name}
+            """, {
+                'object_schema': object_schema,
+                'object_name': object_name,
+            })
 
         return st.rowcount() > 0
 
     def list_schemas(self, schema_name_pattern='%'):
-        st = self._execute_snapshot("""
+        st = self.execute_snapshot("""
             SELECT *
             FROM sys.exa_schemas
             WHERE schema_name LIKE {schema_name_pattern}
@@ -87,7 +112,7 @@ class ExaMetaData(object):
         return st.fetchall()
 
     def list_tables(self, table_schema_pattern='%', table_name_pattern='%'):
-        st = self._execute_snapshot("""
+        st = self.execute_snapshot("""
             SELECT *
             FROM sys.exa_all_tables
             WHERE table_schema LIKE {table_schema_pattern}
@@ -101,7 +126,7 @@ class ExaMetaData(object):
         return st.fetchall()
 
     def list_views(self, view_schema_pattern='%', view_name_pattern='%'):
-        st = self._execute_snapshot("""
+        st = self.execute_snapshot("""
             SELECT *
             FROM sys.exa_all_views
             WHERE view_schema LIKE {view_schema_pattern}
@@ -116,7 +141,7 @@ class ExaMetaData(object):
 
     def list_columns(self, column_schema_pattern='%', column_table_pattern='%'
                      , column_object_type_pattern='%', column_name_pattern='%'):
-        st = self._execute_snapshot("""
+        st = self.execute_snapshot("""
             SELECT *
             FROM sys.exa_all_columns
             WHERE column_schema LIKE {column_schema_pattern}
@@ -133,7 +158,7 @@ class ExaMetaData(object):
         return st.fetchall()
 
     def list_objects(self, object_name_pattern='%', object_type_pattern='%', owner_pattern='%', root_name_pattern='%'):
-        st = self._execute_snapshot("""
+        st = self.execute_snapshot("""
             SELECT *
             FROM sys.exa_all_objects
             WHERE object_name LIKE {object_name_pattern}
@@ -150,7 +175,7 @@ class ExaMetaData(object):
         return st.fetchall()
 
     def list_object_sizes(self, object_name_pattern='%', object_type_pattern='%', owner_pattern='%', root_name_pattern='%'):
-        st = self._execute_snapshot("""
+        st = self.execute_snapshot("""
             SELECT *
             FROM sys.exa_all_object_sizes
             WHERE object_name LIKE {object_name_pattern}
@@ -167,7 +192,7 @@ class ExaMetaData(object):
         return st.fetchall()
 
     def list_indices(self, index_schema_pattern='%', index_table_pattern='%', index_owner_pattern='%'):
-        st = self._execute_snapshot("""
+        st = self.execute_snapshot("""
             SELECT *
             FROM sys.exa_all_indices
             WHERE index_schema LIKE {index_schema_pattern}
@@ -187,18 +212,23 @@ class ExaMetaData(object):
         Never hardcode this list! It might change with next Exasol server version without warning
         """
         if not self.sql_keywords:
-            st = self._execute_snapshot("""
-                SELECT keyword
-                FROM EXA_SQL_KEYWORDS
-                WHERE reserved IS TRUE
-                ORDER BY keyword
-            """)
+            if self.connection.protocol_version() >= constant.PROTOCOL_V2:
+                st = self.execute_meta_nosql('getKeywords')
 
-            self.sql_keywords = st.fetchcol()
+                self.sql_keywords = [r['KEYWORD'] for r in st.fetchall() if r['RESERVED'] is True]
+            else:
+                st = self.execute_snapshot("""
+                    SELECT keyword
+                    FROM EXA_SQL_KEYWORDS
+                    WHERE reserved IS TRUE
+                    ORDER BY keyword
+                """)
+
+                self.sql_keywords = st.fetchcol()
 
         return self.sql_keywords
 
-    def _execute_snapshot(self, query, query_params=None):
+    def execute_snapshot(self, query, query_params=None):
         """
         Execute query in snapshot transaction mode using SQL hint
         fetch_dict=True is enforced to prevent users from relying on order of columns in system views
@@ -207,7 +237,27 @@ class ExaMetaData(object):
             'fetch_dict': True,
         }
 
-        return self.connection.cls_statement(self.connection, f"/*snapshot execution*/{query}", query_params, **options)
+        return self.connection.cls_statement(self.connection, f"{self.snapshot_execution_hint}{query}", query_params, **options)
+
+    def execute_meta_nosql(self, meta_command, meta_params=None):
+        """
+        Execute no SQL meta data command introduced in Exasol 7.0+
+        This feature requires WebSocket protocol v2 or higher
+
+        List of available commands: https://github.com/exasol/websocket-api/blob/master/docs/WebsocketAPIV2.md#metadata-related-commands
+        """
+        if self.connection.protocol_version() < constant.PROTOCOL_V2:
+            raise ExaRuntimeError(self.connection, 'Protocol version 2 is required to execute nosql meta data commands')
+
+        # Security check, prevents execution of dangerous commands if meta_command argument is dynamic
+        if meta_command[0:3] != 'get':
+            raise ExaRuntimeError(self.connection, 'Meta command name should start with prefix \'get*\'')
+
+        options = {
+            'fetch_dict': True,
+        }
+
+        return self.connection.cls_statement(self.connection, meta_command, meta_params, meta_nosql=True, **options)
 
     def __repr__(self):
         return f'<{self.__class__.__name__} session_id={self.connection.session_id()}>'